+++ sqlakit/sql.py
@@ -9,6 +9,7 @@ from ._query import merged
 from ._sql import (
     BaseSQLQuery,
     Context,
+    Inline,
     Param,
     Sql,
     Templates,
@@ -30,6 +31,7 @@ if TYPE_CHECKING:
 __all__ = [
     "SQL",
     "Context",
+    "Inline",
     "Param",
     "SQLQuery",
     "SQLRows",